This is a hack on ConvexHull3d. DON"T USE THIS!
This hack does something like sphere inversion of the points around their midpoint before the hull is generated, then maps the resulting hull vertices back to the originals while maintaining the hull face and edge connectivity. The result is sort of an inverted hull. like a balloon blown up from the centroid instead of shrunk around the set. TODO: I want to do something similar but do a scaled projection on a line instead of using the distance from the centroid.