WIP: Improved inversion center and scaling
This is an improvement on using the centroid of the vertices. It supposes the user will select a ball at the midpoint of the desired hull rather than the centroid of them all.
However, that often means an extra ball must be generated.
TODO: Try using the midpoint of the last 2 selected balls as the center of inversion.
SelectByDiameter with a standard convex hull may still be a better solution.